Adds ability to load system selinux library
That is the first version that tries to locate system selinux library
and loads its.

Adding `selinux` as a dependency to your project should now enable
you to use isolated virtualenvs which previously were giving:

ModuleNotFoundError: No module named 'selinux''

Please note that this will not save your from failing later if
the library does not exist at all on the system.