Check if python bindings are installed
If the virtual-env python version does not match the python version that
contains the real selinux module, the shim will just import itself again
instead of fail. This happens for example when python36 and python3
(which is python37) are installed on Fedora 29 and a virtualenv for
python36 with the selinux shim is used.

Signed-off-by: Till Maas <opensource@till.name>